SDET-U3 Перенес DataProvider в AutorizationTest, оъединил 2 теста
