+++ src/main/resources/aplication.properties
@@ -1,20 +1,16 @@
  username = angular
  password = password
  username2 = tims
-
- username2(1) = smit
- username2(2) = agent
+ username3 = tims
 
  homeMessage = Home
  loginInMessage = You're logged in!!
 
  invalidUsername = Timur
  invalidUsername2 = Bek
- invalidUsername3 = Mat
 
  invalidPassword = 2023
  invalidPassword2 = 123456
- invalidPassword3 = 54321
 
  errorMessage = Username or password is incorrect
